requirements.txt
Flask
pyautogui
pygetwindow
qrcode
Pillow
mss
pyzbar